backend/requirements.txt
# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# CORS
fastapi-cors==0.0.6

# Environment
python-dotenv==1.0.0

# OpenAI
openai==1.12.0

# Vector Database
chromadb==0.4.22

# Confluence API
atlassian-python-api==3.41.0
requests==2.31.0
beautifulsoup4==4.12.3

# PII Detection
presidio-analyzer==2.2.351
presidio-anonymizer==2.2.351
spacy==3.7.2

# Text Processing
langchain==0.1.5
langchain-openai==0.0.5
tiktoken==0.5.2

# Scheduling
apscheduler==3.10.4

# Database (for chat history)
sqlalchemy==2.0.25
aiosqlite==0.19.0

# Utilities
pydantic==2.5.3
pydantic-settings==2.1.0
httpx==0.26.0
